requirements.txt
# for dh
pynacl==1.5.0
cryptography==45.0.3